requirements.txt
# Django核心框架
Django==4.2.14
asgiref==3.9.1
sqlparse==0.5.3

# Django REST Framework - API开发
djangorestframework==3.14.0
django-cors-headers==4.3.1

# 数据库相关 - SQLite FTS5全文检索（Django内置支持）
sqlite-web==0.6.0

# MinIO对象存储
minio==7.2.0
boto3==1.40.2
botocore==1.40.2

# 文本嵌入和机器学习相关
sentence-transformers==5.1.0
torch==2.8.0
numpy==2.3.2
transformers==4.55.2
tokenizers==0.21.4
safetensors==0.6.2
huggingface-hub==0.34.3
hf-xet==1.1.7

# 机器学习工具
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.1
threadpoolctl==3.6.0

# 数据处理和工具
networkx==3.3
sympy==1.13.3
mpmath==1.3.0
packaging==25.0
setuptools==80.9.0
wheel==0.45.1

# 图像处理
pillow==11.3.0

# 网络和HTTP
requests==2.32.4
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10

# 文件系统
fsspec==2025.7.0
filelock==3.18.0

# 文本处理
regex==2025.7.34
PyYAML==6.0.2
tqdm==4.67.1

# 模板引擎
Jinja2==3.1.4
MarkupSafe==2.1.5

# 类型提示
typing_extensions==4.14.1

# 生产环境部署
gunicorn==23.0.0

# 其他工具
six==1.17.0
